pyproject.toml
[project]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Dikshitharyana",email = "dikshitkumar951@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "notebook (>=7.4.3,<8.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.19,<0.4.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "langchain-together (>=0.3.0,<0.4.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "streamlit (>=1.45.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
